Relationships

+++ database/migrations/2019_06_03_162811_create_likes_table.php
@@ -15,8 +15,8 @@ class CreateLikesTable extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('reply_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('reply_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
         });
     }
